keeper-bot/keeper.py
"""
Keeper Bot - Automated Arbitrage Execution
Monitors opportunities and executes trades via smart contracts
"""
import asyncio
import os
from datetime import datetime
from typing import List, Dict, Optional
import httpx
from dotenv import load_dotenv

load_dotenv()

# Configuration
API_BASE_URL = os.getenv("API_BASE_URL", "http://localhost:8001")
MIN_PROFIT_PERCENT = float(os.getenv("MIN_PROFIT_PERCENT", "5.0"))
MAX_TRADE_SIZE = float(os.getenv("MAX_TRADE_SIZE", "1000"))
CHECK_INTERVAL = int(os.getenv("CHECK_INTERVAL", "30"))  # seconds

# Web3 Configuration (TODO: Implement)
VAULT_CONTRACT_ADDRESS = os.getenv("VAULT_CONTRACT_ADDRESS", "")
KEEPER_PRIVATE_KEY = os.getenv("KEEPER_PRIVATE_KEY", "")
RPC_URL = os.getenv("RPC_URL", "https://mainnet.base.org")


class ArbitrageOpportunity:
    """Represents an arbitrage opportunity"""
    def __init__(self, data: Dict):
        self.id = data.get('id')
        self.title = data.get('title')
        self.buy_platform = data.get('buyPlatform')
        self.sell_platform = data.get('sellPlatform')
        self.buy_price = data.get('buyPrice', 0)
        self.sell_price = data.get('sellPrice', 0)
        self.spread_percent = data.get('spreadPercent', 0)
        self.min_volume = data.get('minSideVolume', 0)
        self.estimated_slippage = data.get('estimatedSlippage', 0)
        self.feasibility_score = data.get('feasibilityScore', 0)


class KeeperBot:
    """Main keeper bot for monitoring and executing arbitrage"""
    
    def __init__(self):
        self.client = httpx.AsyncClient(timeout=30.0)
        self.running = False
        self.execution_count = 0
        
    async def fetch_opportunities(self) -> List[ArbitrageOpportunity]:
        """Fetch arbitrage opportunities from API"""
        try:
            response = await self.client.get(f"{API_BASE_URL}/api/arbitrage/opportunities")
            response.raise_for_status()
            data = response.json()
            
            opportunities = []
            for opp_data in data.get('opportunities', []):
                opp = ArbitrageOpportunity(opp_data)
                opportunities.append(opp)
            
            return opportunities
        except Exception as e:
            print(f"Error fetching opportunities: {e}")
            return []
    
    def filter_opportunities(self, opportunities: List[ArbitrageOpportunity]) -> List[ArbitrageOpportunity]:
        """Filter opportunities based on criteria"""
        filtered = []
        
        for opp in opportunities:
            # Check minimum profit threshold
            if opp.spread_percent < MIN_PROFIT_PERCENT:
                continue
            
            # Check feasibility
            if opp.feasibility_score < 50:
                continue
            
            # Check slippage
            if opp.estimated_slippage > 2.0:
                continue
            
            # Check minimum volume
            if opp.min_volume < 100:
                continue
            
            filtered.append(opp)
        
        return filtered
    
    async def execute_arbitrage(self, opportunity: ArbitrageOpportunity) -> bool:
        """Execute arbitrage trade via smart contract"""
        try:
            print(f"\n🎯 Executing arbitrage opportunity:")
            print(f"   Market: {opportunity.title}")
            print(f"   Route: {opportunity.buy_platform} → {opportunity.sell_platform}")
            print(f"   Spread: {opportunity.spread_percent:.2f}%")
            print(f"   Trade Size: ${MAX_TRADE_SIZE}")
            
            # TODO: Implement Web3 contract call
            # 1. Connect to wallet
            # 2. Approve vault contract
            # 3. Call executeArbitrage() on vault contract
            # 4. Wait for confirmation
            # 5. Return result
            
            # For now, simulate execution
            await asyncio.sleep(2)
            
            self.execution_count += 1
            print(f"   ✅ Execution successful (simulated)")
            print(f"   Total executions: {self.execution_count}")
            
            return True
            
        except Exception as e:
            print(f"   ❌ Execution failed: {e}")
            return False
    
    async def monitor_loop(self):
        """Main monitoring loop"""
        print(f"🤖 Keeper Bot Started")
        print(f"   API: {API_BASE_URL}")
        print(f"   Min Profit: {MIN_PROFIT_PERCENT}%")
        print(f"   Max Trade Size: ${MAX_TRADE_SIZE}")
        print(f"   Check Interval: {CHECK_INTERVAL}s")
        print(f"\n{'='*50}\n")
        
        self.running = True
        
        while self.running:
            try:
                # Fetch opportunities
                print(f"[{datetime.now().strftime('%H:%M:%S')}] Checking for opportunities...")
                opportunities = await self.fetch_opportunities()
                
                if not opportunities:
                    print("   No opportunities found")
                else:
                    print(f"   Found {len(opportunities)} opportunities")
                    
                    # Filter opportunities
                    filtered = self.filter_opportunities(opportunities)
                    print(f"   {len(filtered)} opportunities meet criteria")
                    
                    # Execute filtered opportunities
                    for opp in filtered:
                        await self.execute_arbitrage(opp)
                        # Add delay between executions
                        await asyncio.sleep(5)
                
                # Wait before next check
                await asyncio.sleep(CHECK_INTERVAL)
                
            except KeyboardInterrupt:
                print("\n\n🛑 Keeper Bot Stopped")
                self.running = False
                break
            except Exception as e:
                print(f"Error in monitor loop: {e}")
                await asyncio.sleep(CHECK_INTERVAL)
        
        await self.client.aclose()
    
    def stop(self):
        """Stop the keeper bot"""
        self.running = False


async def main():
    """Main entry point"""
    keeper = KeeperBot()
    
    try:
        await keeper.monitor_loop()
    except KeyboardInterrupt:
        print("\nShutting down...")
        keeper.stop()


if __name__ == "__main__":
    asyncio.run(main())
