keeper-bot/requirements.txt
httpx>=0.27.0
python-dotenv>=1.0.0
web3>=6.0.0
asyncio>=3.4.3
